% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/region.R
\name{data}
\alias{data}
\title{Simulate Data for Survival Analysis}
\usage{
data(n, A)

data(n, A)
}
\arguments{
\item{n}{Number of observations}

\item{A}{A matrix representing the region for spatial covariates}
}
\value{
A matrix of simulated data with covariates, time-to-event, and censoring information

A matrix of simulated data with covariates, time-to-event, and censoring information
}
\description{
This function generates simulated data for survival analysis, including covariates,
time-to-event, and censoring times.

This function generates simulated data for survival analysis, including covariates,
time-to-event, and censoring times.
}
