% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{gelman_diagnostics}
\alias{gelman_diagnostics}
\title{Gelman-Rubin Diagnostics for MCMC Convergence}
\usage{
gelman_diagnostics(chain1, chain2)
}
\arguments{
\item{chain1}{MCMC trace from the first chain.}

\item{chain2}{MCMC trace from the second chain.}
}
\value{
A list with diagnostics and plots.
}
\description{
This function performs Gelman-Rubin diagnostics on two MCMC chains.
It provides convergence diagnostics using the Gelman-Rubin statistic.
}
