% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation _ Metropolis-Hastings.R
\name{mcmc_thin}
\alias{mcmc_thin}
\title{Thin the MCMC Trace}
\usage{
mcmc_thin(mcmc_r)
}
\arguments{
\item{mcmc_r}{The MCMC result object.}
}
\value{
A thinned MCMC trace.
}
\description{
This function thins the MCMC trace by selecting one sample for every specified interval.
}
