% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation _ Metropolis-Hastings.R
\name{MSE}
\alias{MSE}
\title{Compute the Mean Squared Error (MSE)}
\usage{
MSE(estimate, true_value)
}
\arguments{
\item{estimate}{Estimated parameter value.}

\item{true_value}{True parameter value.}
}
\value{
The computed MSE.
}
\description{
This function computes the Mean Squared Error (MSE) between estimated and true parameter values.
}
