% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition Spatial Data}
\usage{
partition(A)
}
\arguments{
\item{A}{A matrix representing the region for spatial covariates}
}
\value{
A list containing partition sizes and grid locations
}
\description{
This function divides a given spatial region into smaller partitions.
}
