% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_beta.R
\name{prior_beta}
\alias{prior_beta}
\title{Prior Distribution for Beta Parameters}
\usage{
prior_beta(pars)
}
\arguments{
\item{pars}{A vector of beta parameters}
}
\value{
The sum of the log of the normal prior density for each beta
}
\description{
This function computes the prior distribution for the beta parameters in a survival model.
}
