% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation _ Metropolis-Hastings.R, R/mcmc.R
\name{prog2}
\alias{prog2}
\title{Metropolis-Hastings MCMC for STSU Weibul Model}
\usage{
prog2(A, beta_ini, y, iterations, burn_in)

prog2(A, beta_ini, y, iterations, burn_in)
}
\arguments{
\item{A}{A matrix or data frame for your dataset.}

\item{beta_ini}{Initial parameter values for the MCMC.}

\item{y}{The vector of observed data.}

\item{iterations}{Number of MCMC iterations.}

\item{burn_in}{Number of burn-in iterations to discard.}
}
\value{
An object of class `mcmc` containing the trace of MCMC samples.

An object of class `mcmc` containing the trace of MCMC samples.
}
\description{
This function runs the Metropolis-Hastings MCMC algorithm for fitting the STSU Weibul model.

A wrapper function to apply the Metropolis-Hastings algorithm for MCMC.
}
