% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Generate a Region}
\usage{
region(n, seed)
}
\arguments{
\item{n}{Number of points to generate.}

\item{seed}{Seed for random number generation.}
}
\value{
A matrix of points.
}
\description{
This function generates a region of random points and plots them.
}
