% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation _ Metropolis-Hastings.R
\name{run_mcmc_loop}
\alias{run_mcmc_loop}
\title{Run MCMC Loop for Model Fitting}
\usage{
run_mcmc_loop(A, y, iterations = 10000, burn_in = 1000, a = 50)
}
\arguments{
\item{A}{The data matrix for the model.}

\item{y}{The observed data.}

\item{iterations}{The number of MCMC iterations.}

\item{burn_in}{The number of burn-in iterations.}
}
\value{
A list of calculated metrics (DIC, BIC, pD, etc.).
}
\description{
This function runs the MCMC loop for a specified number of iterations and calculates various diagnostics (DIC, BIC, etc.).
}
