% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stsu_likelihood.R
\name{stsu_likelihood}
\alias{stsu_likelihood}
\title{Likelihood for STSU Weibul Model}
\usage{
stsu_likelihood(pars, y, nL, zA)
}
\arguments{
\item{pars}{A vector of parameters for the Weibul model.}

\item{y}{A vector of observed event times.}

\item{nL}{A vector of partition sizes.}

\item{zA}{A vector of normalized partition weights.}
}
\value{
The likelihood value for the STSU Weibul model.
}
\description{
This function computes the likelihood for the STSU Weibul model using the provided parameters.
}
