% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zA_vector.R
\name{zA_vector}
\alias{zA_vector}
\title{Calculate zA and nL Vectors}
\usage{
zA_vector(n, A)
}
\arguments{
\item{n}{The number of data points.}

\item{A}{A matrix or data frame representing spatial data points.}
}
\value{
A list containing the partition sizes `nL` and the normalized `zA` vector.
}
\description{
This function calculates the `zA` vector and partition sizes (`nL`) based on input data `A`.
}
